Add solution for Exercise Collections Enums
