+++ exercise/g_collections_enums/src/main.rs
@@ -1,77 +1,21 @@
-// Silence some warnings that could distract from the exercise
-#![allow(unused_variables, unused_mut, dead_code)]
-
-// Someone is shooting arrows at a target.  We need to classify the shots.
-//
-// 1a. Create an enum called `Shot` with variants:
-// - `Bullseye`
-// - `Hit`, containing the distance from the center (an f64)
-// - `Miss`
-//
-// You will need to complete 1b as well before you will be able to run this program successfully.
-
-impl Shot {
-    // Here is a method for the `Shot` enum you just defined.
-    fn points(self) -> i32 {
-        // 1b. Implement this method to convert a Shot into points
-        // - return 5 points if `self` is a `Shot::Bullseye`
-        // - return 2 points if `self` is a `Shot::Hit(x)` where x < 3.0
-        // - return 1 point if `self` is a `Shot::Hit(x)` where x >= 3.0
-        // - return 0 points if `self` is a Miss
-    }
-}
+use g_collections_enums::{get_arrow_coords, Coord, Shot};
 
 fn main() {
     // Simulate shooting a bunch of arrows and gathering their coordinates on the target.
     let arrow_coords: Vec<Coord> = get_arrow_coords(5);
-    let mut shots: Vec<Shot> = Vec::new();
+    let shots: Vec<Shot> = arrow_coords
+        .iter()
+        .map(|coord| {
+            coord.print_description();
+            match coord.distance_from_center() {
+                x if x < 1.0 => Shot::Bullseye,
+                x if x <= 5.0 => Shot::Hit(x),
+                _ => Shot::Miss,
+            }
+        })
+        .collect();
 
-    // 2. For each coord in arrow_coords:
-    //
-    //   A. Call `coord.print_description()`
-    //   B. Append the correct variant of `Shot` to the `shots` vector depending on the value of
-    //   `coord.distance_from_center()`
-    //      - Less than 1.0 -- `Shot::Bullseye`
-    //      - Between 1.0 and 5.0 -- `Shot::Hit(value)`
-    //      - Greater than 5.0 -- `Shot::Miss`
-
-
-    let mut total = 0;
-    // 3. Finally, loop through each shot in shots and add its points to total
+    let total: i32 = shots.iter().map(|shot| shot.points()).sum();
 
     println!("Final point total is: {}", total);
-}
-
-// A coordinate of where an Arrow hit
-#[derive(Debug)]
-struct Coord {
-    x: f64,
-    y: f64,
-}
-
-impl Coord {
-    fn distance_from_center(&self) -> f64 {
-        (self.x.powf(2.0) + self.y.powf(2.0)).sqrt()
-    }
-    fn print_description(&self) {
-        println!(
-            "coord is {:.1} away, at ({:.1}, {:.1})",
-            self.distance_from_center(),
-            self.x,
-            self.y);
-    }
-
-}
-
-// Generate some random coordinates
-fn get_arrow_coords(num: u32) -> Vec<Coord> {
-    let mut coords: Vec<Coord> = Vec::new();
-    for _ in 0..num {
-        let coord = Coord {
-            x: (rand::random::<f64>() - 0.5) * 12.0,
-            y: (rand::random::<f64>() - 0.5) * 12.0,
-        };
-        coords.push(coord);
-    }
-    coords
 }
